fix(Controller): 기존 Controller 어노테이션 기반으로 변경
