--- app/src/main/java/camp/nextstep/controller/RegisterViewController.java
@@ -1,14 +1,0 @@
-package camp.nextstep.controller;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class RegisterViewController implements Controller {
-
-    @Override
-    public ModelAndView execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        return ModelAndView.ofJspView("/register.jsp");
-    }
-}
